Move supporting/base web files to http subdirectory.
