+++ Makefile.am
@@ -52,15 +52,15 @@ src_libbitcoin_server_la_SOURCES = \
     src/services/transaction_service.cpp \
     src/web/block_socket.cpp \
     src/web/heartbeat_socket.cpp \
-    src/web/json_string.cpp \
     src/web/query_socket.cpp \
-    src/web/socket.cpp \
     src/web/transaction_socket.cpp \
     src/web/http/connection.cpp \
     src/web/http/connection.hpp \
     src/web/http/http.hpp \
+    src/web/http/json_string.cpp \
     src/web/http/manager.cpp \
     src/web/http/manager.hpp \
+    src/web/http/socket.cpp \
     src/web/http/utilities.cpp \
     src/web/http/utilities.hpp \
     src/workers/authenticator.cpp \
@@ -136,10 +136,13 @@ include_bitcoin_server_webdir = ${includedir}/bitcoin/server/web
 include_bitcoin_server_web_HEADERS = \
     include/bitcoin/server/web/block_socket.hpp \
     include/bitcoin/server/web/heartbeat_socket.hpp \
-    include/bitcoin/server/web/json_string.hpp \
     include/bitcoin/server/web/query_socket.hpp \
-    include/bitcoin/server/web/socket.hpp \
     include/bitcoin/server/web/transaction_socket.hpp
+
+include_bitcoin_server_web_httpdir = ${includedir}/bitcoin/server/web/http
+include_bitcoin_server_web_http_HEADERS = \
+    include/bitcoin/server/web/http/json_string.hpp \
+    include/bitcoin/server/web/http/socket.hpp
 
 include_bitcoin_server_workersdir = ${includedir}/bitcoin/server/workers
 include_bitcoin_server_workers_HEADERS = \
