+++ include/bitcoin/server/web/http/json_string.hpp
