+++ include/bitcoin/server/web/http/socket.hpp
