+++ include/bitcoin/server/web/query_socket.hpp
@@ -24,7 +24,7 @@
 #include <bitcoin/protocol.hpp>
 #include <bitcoin/server/define.hpp>
 #include <bitcoin/server/settings.hpp>
-#include <bitcoin/server/web/socket.hpp>
+#include <bitcoin/server/web/http/socket.hpp>
 
 namespace libbitcoin {
 namespace server {
