+++ src/web/heartbeat_socket.cpp
@@ -23,7 +23,7 @@
 #include <bitcoin/protocol.hpp>
 #include <bitcoin/server/server_node.hpp>
 #include <bitcoin/server/settings.hpp>
-#include <bitcoin/server/web/json_string.hpp>
+#include <bitcoin/server/web/http/json_string.hpp>
 
 namespace libbitcoin {
 namespace server {
